docs: iperf3.md: add in more steps for ensuring IP address are correctly set
We need to use 'sudo ifdown <adapter devname>' to bring down adapter device
and 'sudo ifup <adapter devname> to fire it up so that it takes effect of
the setting in /etc/network/interfaces.

For server-end:
We remove 'netmask' and 'gateway' settings from /etc/network/interfaces as
they are discoverable through DHCP. Setting the 'gateway' value for enp0s8
will cause confusion to Linux IP route table as the link is not meant for
Internet access and DUT is not serving as gateway.

For enp0s3 which are the uplink port (for Internet), we remove all setting
of DNS server since these can be discoverable through DHCP.

For DUT (client-end):
We remove 'gateway' setting from /etc/network/interfaces as they are not
required since device is not supposed to allow to access Internet.

Signed-off-by: Ong Boon Leong <elvinongbl@gmail.com>